requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
exceptiongroup==1.0.0rc9
fake-useragent==0.1.11
h11==0.14.0
html5lib==1.1
idna==3.4
lxml==4.9.1
numpy==1.23.4
outcome==1.2.0
pandas==1.5.0
py==1.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
retry==0.9.2
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webencodings==0.5.1
wsproto==1.2.0
XlsxWriter==3.0.3
